Chat Application for Android Devices
This app is used to communicate from one device to another device using
sockets. The main purpose of this app is to communicate with others who
are connected in same network. (For example: Manger can act as server
and broadcast message to all this employees who are connected to the
same socket number.)